support 64 bit

+++ apps/manager/src/manager.c
@@ -161,7 +161,7 @@ static void init_env_colour(m_env_t *env) {
     color_allocator_t *init_allocator; 
     color_allocator_t *color_allocator[CC_NUM_DOMAINS]; 
     reservation_t v_reserve; 
-    uint32_t div[CC_NUM_DOMAINS] = {CC_DIV, CC_DIV}; 
+    size_t div[CC_NUM_DOMAINS] = {CC_DIV, CC_DIV}; 
     int error = 0; 
     void *vaddr;
